refactor(compare): collapse saved-compare detail into the report page
The saved-comparison "detail" (`/benchmarks/compare/saved/:id`,
`SavedComparePage`) and the standalone report viewer (`/reports/:id`,
`ReportPage`) were two pages for one thing: the detail page embedded the
narrative report and then appended a raw `ReportSections` block (Test
matrix + Key metrics + bar charts + environment) *after* the report's own
footer — reading as "report's done… here come more tables".

Consolidate onto `/reports/:id`:

- Delete `SavedComparePage`; the list cards and the save-and-generate
  dialog now navigate straight to `/reports/:id` (`?generate=1` bridge and
  generate/regenerate/export/delete all move into the report page's slim
  top bar, which is hidden in print). Empty state generates inline instead
  of bouncing to a detail page.
- Drop the raw Key metrics / bar charts / environment from the report
  surface. Keep only a lightweight "data source" table inside
  `SavedCompareReport`, rendered before the footer so the footer truly
  closes the document. Identity columns only (stage / name / tool /
  scenario); names deep-link to the benchmark detail.
- `ReportSections` / `CompareGrid` / `StageBarChartsSection` stay — they're
  still used by `BenchmarkComparePage` (ad-hoc compare).
