+++ apps/web/src/features/benchmarks/compare/ReportPage.tsx
@@ -1,15 +1,28 @@
 import type { Benchmark, CompareNarrative } from "@modeldoctor/contracts";
-import { ArrowLeft, Printer } from "lucide-react";
+import { ArrowLeft, Download, Printer, RefreshCw, Sparkles, Trash2 } from "lucide-react";
+import { useCallback, useEffect, useRef, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { Link, useParams } from "react-router-dom";
+import { Link, useNavigate, useParams, useSearchParams } from "react-router-dom";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
 import { Button } from "@/components/ui/button";
-import { useSavedCompare } from "./queries";
-import { type ReportRun, type ReportSectionsProps } from "./ReportSections";
+import { useLlmJudgeProvider } from "@/features/settings/queries";
+import { exportPageAsHtml } from "./exportHtml";
+import { useDeleteSavedCompare, useSavedCompare, useSynthesizeSavedCompare } from "./queries";
+import { ReportProgress } from "./ReportProgress";
+import type { ReportRun } from "./ReportSections";
 import { SavedCompareReport } from "./SavedCompareReport";
 
-function extractParamsSummary(
-  params: unknown,
-): ReportSectionsProps["runs"][number]["paramsSummary"] {
+function extractParamsSummary(params: unknown): ReportRun["paramsSummary"] {
   if (!params || typeof params !== "object") return {};
   const p = params as Record<string, unknown>;
   return {
@@ -20,22 +33,67 @@ function extractParamsSummary(
 }
 
 /**
- * Standalone report viewer route: `/reports/:id`.
+ * Single saved-comparison report page: `/reports/:id`.
  *
- * Sits outside `<AppShell>` so the report takes the full viewport width
- * without the sidebar or app chrome competing for space. Its only non-report
- * UI is a slim top bar with brand + back-to-app + print button. The top bar
- * is hidden under `@media print` so PDF output starts straight at the Hero.
+ * Sits outside `<AppShell>` so the report takes the full viewport width without
+ * the sidebar competing for space. It is BOTH the management surface (generate /
+ * regenerate / export / delete, in the slim top bar) and the reading/print
+ * surface — there is no separate detail page. The top bar is hidden under
+ * `@media print` so PDF output starts straight at the Hero.
  *
- * Use cases:
- *   - Read the AI-generated deep report
- *   - Share the URL with partner / client (still gated by JWT auth)
- *   - Print to PDF via window.print() (Primer print stylesheet kicks in)
+ * States:
+ *   - loading / not-found
+ *   - no narrative yet → inline "Generate report" card (auto-fires when the
+ *     save-and-generate flow lands here with `?generate=1`)
+ *   - narrative present → the Primer-styled report + a "Data source" table
  */
 export function ReportPage() {
   const { id = "" } = useParams<{ id: string }>();
   const { t } = useTranslation("benchmarks");
+  const navigate = useNavigate();
   const query = useSavedCompare(id);
+  const provider = useLlmJudgeProvider();
+  const synth = useSynthesizeSavedCompare(id);
+  const del = useDeleteSavedCompare();
+  const [narrativeOverride, setNarrativeOverride] = useState<CompareNarrative | null>(null);
+  const [searchParams, setSearchParams] = useSearchParams();
+  const generateParam = searchParams.get("generate");
+  const autoGenFired = useRef(false);
+
+  const generate = useCallback(async () => {
+    const r = await synth.mutateAsync({ locale: "zh-CN" });
+    setNarrativeOverride(r.narrative);
+  }, [synth.mutateAsync]);
+
+  // Depend on primitives, not the whole query.data / searchParams objects, so
+  // background refetches and unrelated URL changes don't re-run this effect.
+  const scId = query.data?.id;
+  const hasNarrative = !!query.data?.narrative;
+
+  // Save-and-generate bridge: SaveCompareDialog navigates here with ?generate=1
+  // after a save-and-generate. Fire synthesize once, then strip the flag.
+  useEffect(() => {
+    if (autoGenFired.current) return;
+    if (generateParam !== "1") return;
+    if (!scId) return; // data not loaded yet — wait, don't strip
+    if (hasNarrative) {
+      autoGenFired.current = true;
+      setSearchParams({}, { replace: true });
+      return;
+    }
+    if (!provider.data?.enabled || synth.isPending) return;
+    autoGenFired.current = true;
+    setSearchParams({}, { replace: true });
+    void generate();
+  }, [
+    generateParam,
+    scId,
+    hasNarrative,
+    provider.data?.enabled,
+    synth.isPending,
+    setSearchParams,
+    generate,
+  ]);
 
   if (query.isLoading) {
     return (
@@ -60,26 +118,7 @@ export function ReportPage() {
   }
 
   const sc = query.data;
-  const narrative = sc.narrative as CompareNarrative | null;
-
-  if (!narrative) {
-    return (
-      <div className="flex min-h-screen flex-col items-center justify-center gap-3">
-        <p className="text-sm text-muted-foreground">
-          {t("savedCompare.reportPage.notGenerated", {
-            defaultValue: "This comparison does not have an AI report yet.",
-          })}
-        </p>
-        <Button asChild>
-          <Link to={`/benchmarks/compare/saved/${sc.id}`}>
-            {t("savedCompare.reportPage.openDetailToGenerate", {
-              defaultValue: "Open detail page to generate",
-            })}
-          </Link>
-        </Button>
-      </div>
-    );
-  }
+  const narrative = narrativeOverride ?? (sc.narrative as CompareNarrative | null);
 
   const reportRuns: ReportRun[] = sc.benchmarks.map((b) => ({
     id: b.id,
@@ -100,20 +139,32 @@ export function ReportPage() {
     paramsSummary: extractParamsSummary(b.params),
   }));
 
+  async function onDelete() {
+    await del.mutateAsync(id);
+    navigate("/benchmarks/compare/saved");
+  }
+
+  function onExport() {
+    const root = document.querySelector("[data-report-root]") as HTMLElement | null;
+    if (root) void exportPageAsHtml(root, sc.name);
+  }
+
   function onPrint() {
     window.print();
   }
 
+  const canGenerate = !!provider.data?.enabled && !synth.isPending;
+
   return (
     <div className="min-h-screen bg-[var(--pr-bg-subtle,#f6f8fa)]">
-      {/* Slim brand bar — hidden in print */}
+      {/* Slim management/brand bar — hidden in print */}
       <header className="pr-app-bar sticky top-0 z-50 border-b border-border bg-background/90 backdrop-blur-sm">
         <div className="mx-auto flex max-w-[1760px] items-center justify-between gap-4 px-8 py-3">
           <div className="flex items-center gap-3">
             <Button variant="ghost" size="sm" asChild>
-              <Link to={`/benchmarks/compare/saved/${sc.id}`}>
+              <Link to="/benchmarks/compare/saved">
                 <ArrowLeft className="mr-1 h-4 w-4" />
-                {t("savedCompare.reportPage.backToDetail", { defaultValue: "Back to detail" })}
+                {t("savedCompare.reportPage.backToList", { defaultValue: "Back to comparisons" })}
               </Link>
             </Button>
             <div className="flex items-center gap-2 text-sm text-muted-foreground">
@@ -123,19 +174,100 @@ export function ReportPage() {
             </div>
           </div>
           <div className="flex items-center gap-2">
-            <Button variant="outline" size="sm" onClick={onPrint}>
-              <Printer className="mr-1.5 h-4 w-4" />
-              {t("savedCompare.reportPage.printPdf", { defaultValue: "Print / PDF" })}
-            </Button>
+            {narrative ? (
+              <>
+                <Button
+                  variant="outline"
+                  size="sm"
+                  onClick={() => void generate()}
+                  disabled={!canGenerate}
+                >
+                  <RefreshCw
+                    className={`mr-1.5 h-4 w-4 ${synth.isPending ? "animate-spin" : ""}`}
+                  />
+                  {t("savedCompare.detail.regenerate")}
+                </Button>
+                <Button variant="outline" size="sm" onClick={onExport}>
+                  <Download className="mr-1.5 h-4 w-4" />
+                  {t("savedCompare.detail.export")}
+                </Button>
+                <Button variant="outline" size="sm" onClick={onPrint}>
+                  <Printer className="mr-1.5 h-4 w-4" />
+                  {t("savedCompare.reportPage.printPdf", { defaultValue: "Print / PDF" })}
+                </Button>
+              </>
+            ) : null}
+            <AlertDialog>
+              <AlertDialogTrigger asChild>
+                <Button
+                  variant="ghost"
+                  size="sm"
+                  className="text-destructive hover:text-destructive"
+                >
+                  <Trash2 className="mr-1.5 h-4 w-4" />
+                  {t("savedCompare.detail.deleteTitle")}
+                </Button>
+              </AlertDialogTrigger>
+              <AlertDialogContent>
+                <AlertDialogHeader>
+                  <AlertDialogTitle>{t("savedCompare.detail.deleteTitle")}</AlertDialogTitle>
+                  <AlertDialogDescription>
+                    {t("savedCompare.detail.deleteBody")}
+                  </AlertDialogDescription>
+                </AlertDialogHeader>
+                <AlertDialogFooter>
+                  <AlertDialogCancel>{t("savedCompare.dialog.cancel")}</AlertDialogCancel>
+                  <AlertDialogAction onClick={onDelete}>
+                    {t("savedCompare.detail.deleteTitle")}
+                  </AlertDialogAction>
+                </AlertDialogFooter>
+              </AlertDialogContent>
+            </AlertDialog>
           </div>
         </div>
+        {synth.isPending ? (
+          <div className="mx-auto max-w-[1760px] px-8 pb-3">
+            <ReportProgress active={synth.isPending} />
+          </div>
+        ) : null}
       </header>
 
-      <SavedCompareReport
-        narrative={narrative}
-        runs={reportRuns}
-        printHeader={`ModelDoctor · ${sc.name}`}
-      />
+      {narrative ? (
+        <SavedCompareReport
+          narrative={narrative}
+          runs={reportRuns}
+          printHeader={`ModelDoctor · ${sc.name}`}
+        />
+      ) : (
+        <div className="mx-auto flex max-w-2xl flex-col items-center gap-4 px-8 py-24 text-center">
+          <div className="flex items-center gap-2 text-base font-semibold">
+            <Sparkles className="h-5 w-5 text-violet-500" />
+            {t("savedCompare.reportPage.notGenerated", {
+              defaultValue: "This comparison does not have an AI report yet.",
+            })}
+          </div>
+          <p className="text-sm text-muted-foreground">
+            {t("savedCompare.report.generateHint", {
+              defaultValue: "60-180s deep report — Hero + 6 sections + summary cards (zh-CN).",
+            })}
+          </p>
+          {synth.error ? <p className="text-sm text-destructive">{synth.error.message}</p> : null}
+          <Button onClick={() => void generate()} disabled={!canGenerate}>
+            <Sparkles className="mr-1.5 h-4 w-4" />
+            {synth.isPending
+              ? t("savedCompare.report.generating", { defaultValue: "Generating…" })
+              : t("savedCompare.report.generateButton")}
+          </Button>
+          {!provider.data?.enabled ? (
+            <p className="text-xs text-muted-foreground">
+              {t("savedCompare.report.providerDisabledHint", {
+                defaultValue:
+                  "Configure an LLM judge provider in Settings to enable report generation.",
+              })}
+            </p>
+          ) : null}
+        </div>
+      )}
     </div>
   );
 }
